+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.2] - 15/8/2020
+
+- Added Jiggle Builder
+- Added Builder Extension Methods
+
 ## [0.0.1] - 15/8/2020
 
 - Initial API